refactor: decouple background job from its registration
